Refactoring, take the run-loop out of Task.scala
